requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
black==23.12.1
click==8.1.7
colorama==0.4.6
docopt==0.6.2
fastapi==0.109.2
fastapi-auth0==0.5.0
pipreqs
#psycopg2-binary==2.9.9
psycopg2==2.9.9
fastapi-pagination==0.12.14
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
fastapi-limiter==0.1.6
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
iniconfig==2.0.0
Mako==1.3.0
pathspec==0.12.1
pipreqs==0.4.13
pluggy==1.3.0
pyasn1==0.5.1
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
redis==5.1.0b1
rsa==4.9
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.36.3
uvicorn==0.25.0
yarg==0.1.9
sqlalchemy_utils==0.37.8
itsdangerous==2.1.2
Authlib==1.3.0
modal===0.61.77
boto3~=1.34.36
botocore~=1.34.36
python-multipart==0.0.7
broadcaster==0.2.0
Jinja2==3.1.3
google-cloud-storage
google